Updates Serverspec tests for pip requirements in dev and staging
Recent changes in #1445 bumped a few test pip requirements. This commit
updates the Serverspec tests to match. The Serverspec tests aren't DRY
in this regard; there are two points of update right now, with a few
libs duplicated between `development` and `app-staging`. Not going to
mess with that now, because I just separated them out a few days ago.